requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bidict==0.23.1
click==8.2.0
databases==0.9.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
wsproto==1.2.0
python-jose[cryptography]
passlib
python-jose[cryptography]
passlib
python-multipart
